Elixir Examples v0.0.1 Euler1

A demonstration of potential solutions to the Project Euler 1 in Elixir

Functions

using_calculate_multiples(value)

Specs

using_calculate_multiples(integer) :: integer

Returns the sum of all values below a given number which are multiples of 3 or 5.

While working on Euler2 I realized the calculating the multiples might actually be faster than iterating over every number - which lead to this solution. We calculate the results for multiples of 3 and 5 separately - adding each result to the accumulator.

It turned out to be around 85% faster than the next fastest answer for all examples.

Example

iex> Euler1.using_calculate_multiples(10)
23

using_calculate_multiples(list, value, acc)

using_map_reduce(value)

Specs

using_map_reduce(integer) :: integer

Returns the sum of all values below a given number which are multiples of 3 or 5.

Included while trying to figure out if Enum.sum benefitted from Stream and to show a third solution to the problem. This version uses Stream.filter to map the values and then reduces the result. Oddly - if we ignore using_calculate_multiples/1 - this one seems to perform best at 500,000 items (about 6% faster than using_stream/1) and about the same as the using_stream/1 version at the other levels.

Example

iex> Euler1.using_map_reduce(10)
23

using_stream(value)

Specs

using_stream(integer) :: integer

Returns the sum of all values below a given number which are multiples of 3 or 5.

This solution is a pipeline of transforms ending with a sum. To be honest I wasn’t sure if Stream’s laziness benefits Enum.sum/1 or not… doesn’t seem to however this method was definitely more performant at larger values. It’s worth noting that the private methods were not refactored out when I started the 2nd approach - rather I started out writing using stream as it is and figured out the implementation of the private methods after the fact. I could have avoided the private methods and left the transformation in line - but like the more declarative pipeline in the function serving as the interface. It’s more lines of code but easier to follow.

If we ignore using_calculate_multiples/1 this method is fastest above 100,000 items - except that at 500,000 items using_map_reduce/1 ended up winning.

Example

iex> Euler1.using_stream(10)
23

using_tail_call_optimization(value)

Specs

using_tail_call_optimization(integer) :: integer

Returns the sum of all values below a given number which are multiples of 3 or 5.

Once I finished using_stream/1 I was curious how the method would look using recursion. In this case I used a micropattern - having one arity of a method serve as the interface and another arity of the same method serve as the implementation.

I originally had a “local assignment” here: value_to_add = if desired_value?(head), do: head, else: 0 but considered it a code smell and moved it to it’s own method. Definitely made using_tail_call_optimization/2 easier to read.

If we ignore using_calculate_multiples/1 this method is fastest up to about 100,000 items.

Example

iex> Euler1.using_tail_call_optimization(10)
23