Elixir Examples v0.0.1 Euler2

A demonstration of potential solutions to the Project Euler 2 in Elixir

Functions

perform(top_value)

Specs

perform(integer) :: integer

perform({integer, integer, integer, integer}) ::
  {integer, integer, integer, integer} |
  integer

By considering the terms in the Fibonacci sequence whose values do not exceed a max value, find the sum of the even-valued terms.

We start out by calling perform/1 - which serves as an interface when it’s passed an integer. It creates a tuple to track data with initial settings based on the third Fibonacci number (3)… largely to simplify the methods. It then recurses - using a pipeline to trnasform the tuple until it hits the max value.

Example

iex> Euler2.perform(50)
44